pyproject.toml
[project]
name = "inf367a-image2biomass"
version = "0.1.0"
description = "Project for V2026"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "attrs>=25.4.0",
    "h5py>=3.16.0",
    "loguru>=0.7.3",
    "opencv-python>=4.13.0.90",
    "pandas>=3.0.0",
    "scikit-learn>=1.8.0",
    "tabpfn>=2.0.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]

[tool.setuptools]
package-dir = { "" = "src" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"

[tool.setuptools.packages.find]
where = ["src"]

# Use the cu130 wheels on Linux/Windows (CUDA 13). macOS has no cu130 wheels,
# so it falls back to the default PyPI index and pulls the CPU/MPS build.
[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
